返回结构处理

+++ util-arch/src/main/java/com/wing/java/util/plugin/jdbc/dialect/H2Dialect.java
@@ -3,9 +3,8 @@ package com.wing.java.util.plugin.jdbc.dialect;
 
 /**
  * A dialect compatible with the H2 database.
- * 
- * @author Thomas Mueller
  *
+ * @author Thomas Mueller
  */
 public class H2Dialect extends Dialect {
 
@@ -13,18 +12,18 @@ public class H2Dialect extends Dialect {
         return true;
     }
 
-	public String getLimitString(String sql, int offset,String offsetPlaceholder, int limit, String limitPlaceholder) {
-		return new StringBuffer(sql.length() + 40).
-			append(sql).
-			append((offset > 0) ? " limit "+limitPlaceholder+" offset "+offsetPlaceholder : " limit "+limitPlaceholder).
-			toString();
-	}
+    public String getLimitString(String sql, long offset, String offsetPlaceholder, long limit, String limitPlaceholder) {
+        return new StringBuffer(sql.length() + 40).
+                append(sql).
+                append((offset > 0) ? " limit " + limitPlaceholder + " offset " + offsetPlaceholder : " limit " + limitPlaceholder).
+                toString();
+    }
 
-	@Override
-	public boolean supportsLimitOffset() {
-		return true;
-	}
-    
+    @Override
+    public boolean supportsLimitOffset() {
+        return true;
+    }
+
 //    public boolean bindLimitParametersInReverseOrder() {
 //        return true;
 //    }    
@@ -33,6 +32,5 @@ public class H2Dialect extends Dialect {
 //        return false;
 //    }
 
-    
 
 }
